backend/pyproject.toml
[project]
name = "second-brain"
version = "0.1.0"
description = "AI Second Brain — Pydantic AI agents with Mem0 + Supabase backend."
requires-python = ">=3.11"
dependencies = [
    "pydantic-ai[anthropic]>=0.1.0,<1.0.0",
    "mem0ai>=0.1.0,<1.0.0",
    "supabase>=2.0.0,<3.0.0",
    "python-dotenv>=1.0.0,<2.0.0",
    "pydantic-settings>=2.0.0,<3.0.0",
    "fastmcp>=2.14.0,<3.0.0",
    "click>=8.0.0,<9.0.0",
    "tenacity>=8.0.0,<9.0.0",
    "voyageai>=0.3.0,<1.0.0",
    "Pillow>=10.0.0,<12.0.0",
    "fastapi[standard]>=0.126.0,<0.130.0",
    "httpx>=0.23.0,<0.29.0",
]

[project.optional-dependencies]
ollama = ["pydantic-ai[ollama]>=0.1.0,<1.0.0"]
voyage = ["voyageai>=0.3.0,<1.0.0"]
graphiti = ["graphiti-core[anthropic,falkordb,voyageai]>=0.1.0,<1.0.0"]
subscription = ["claude-agent-sdk>=0.1.0"]
frontend = [
    "streamlit>=1.46.0,<2.0.0",
    "st-link-analysis>=0.4.0",
]
dev = [
    "pydantic-ai[ollama]>=0.1.0,<1.0.0",
    "pytest>=8.0.0,<9.0.0",
    "pytest-asyncio>=0.23.0,<1.0.0",
    "claude-agent-sdk>=0.1.0",
    "httpx>=0.23.0,<0.29.0",
]

[project.scripts]
brain = "second_brain.cli:cli"

[tool.pytest.ini_options]
asyncio_mode = "auto"

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
